+++ xnat_ingest/api/__init__.py
@@ -1,9 +1,9 @@
-from .associate_ import associate
-from .assign_ import assign, INVALID_DIRNAME
-from .check_upload_ import check_upload
-from .group_ import group, group_orthanc, BUILD_NAME_DEFAULT
-from .upload_ import upload
-from .deidentify_ import deidentify
+from .associate_api import associate
+from .assign_api import assign, INVALID_DIRNAME
+from .check_upload_api import check_upload
+from .group_api import group, group_orthanc, BUILD_NAME_DEFAULT
+from .upload_api import upload
+from .deidentify_api import deidentify
 
 __all__ = [
     "upload",
